Server task schedule privileges

+++ app/Http/Controllers/API/ServersTasksController.php
@@ -29,6 +29,8 @@ class ServersTasksController extends AuthController
      */
     public function getList(Server $server)
     {
+        $this->authorize('server-tasks', $server);
+
         return $this->repository->getTasks($server->id);
     }
 
@@ -39,6 +41,13 @@ class ServersTasksController extends AuthController
      */
     public function store(ServerTaskCreateRequest $request)
     {
+        $fields = $request->all();
+
+        $server = Server::findOrFail($fields['server_id']);
+        $this->authorize('server-tasks', $server);
+
+        $this->commandAuthorize($fields['command'], $server);
+
         $serverTaskId = $this->repository->store($request->all());
 
         return response()->json([
@@ -57,7 +66,14 @@ class ServersTasksController extends AuthController
      */
     public function update(ServerTaskUpdateRequest $request, Server $server, ServerTask $serverTask)
     {
-        $this->repository->update($serverTask->id, $request->all());
+        $fields = $request->all();
+
+        $server = Server::findOrFail($fields['server_id']);
+
+        $this->authorize('server-tasks', $server);
+        $this->commandAuthorize($fields['command'], $server);
+
+        $this->repository->update($serverTask->id, $fields);
 
         return response()->json(['message' => 'success'], Response::HTTP_OK);
     }
@@ -71,7 +87,36 @@ class ServersTasksController extends AuthController
      */
     public function destroy(Server $server, ServerTask $serverTask)
     {
+        $this->authorize('server-tasks', $server);
+
         $serverTask->delete();
         return response()->json(['message' => 'success'], Response::HTTP_OK);
+    }
+
+    /**
+     * @param string $command
+     * @param Server $server
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
+    private function commandAuthorize(string $command, Server $server): void
+    {
+        switch ($command) {
+            case 'start':
+                $this->authorize('server-start', $server);
+                break;
+
+            case 'stop':
+                $this->authorize('server-stop', $server);
+                break;
+
+            case 'restart':
+                $this->authorize('server-restart', $server);
+                break;
+
+            case 'update':
+            case 'reinstall':
+                $this->authorize('server-update', $server);
+                break;
+        }
     }
 }
